усложнение вида программы
добавлены папки под распределение модулей
